feat(settings): split up settings page
